Separate frame creation in own submodule
* Places frame creation code in its own /Frame subdirectory
  This so we can create frames alone, which are used in
  property data creation
* Adds code looping property data JSON
* Adds property type structure

+++ AWO/Src/Game/Data/Unit/_units_data.h
@@ -1,10 +1,11 @@
 #pragma once
 
 #include "Game/Data/Unit/units_data.h"
+#include "types.h"
 
 typedef struct Unit_Type_Data
 {
-    int variations_count;    // Amount of variations used by this unit type
+    Uint8 variations_count;  // Amount of variations used by this unit type
     Animation*** variations; // Animation data for every variation of this unit type
 } Unit_Type_Data;
